Fix isset parsing edge cases
Previously, the parser only checked if a selector was bordered on the
right side by a paren, OR operator or AND operator in order to determine
if the selector was an isset operand. This lead to edge cases where
selectors on the right side of binary expressions were incorrectly
getting treated as isset operands.

This commit implements a more robust check for isset operands.

As a result of this work, tests/comparison_filter/046.phpt now works.