Reduce memory allocations
This commit aims to reduce the number of times emalloc() is called.
Instead of allocating arrays and strings each time every time a
literal or node list is compared in the interpreter, create literals and
node list once at parse time.

In addition, use a stack-allocated memory pool for AST nodes. This has
the benefit of eliminating calls to emalloc(), allocating contiguous
blocks of memory, and reducing the risk of future memory leaks.

The tradeoff we're making here is that there is a fixed number of AST
nodes (64), so there's a limit imposed on the size of the JSONPath
query. We'll see in the future if this poses a problem, in which case
we can implement Viktor's dynamic memory pool solution. (6cf186b)