Fix Windows builds
First, we need to compile the ancillary files in the src/ directory.

Then we also need to cater to the local variable `s_lh`, what is
already defined as macro in WinSock2.h.  We could undefine that macro,
but just renaming the variable to `zs_lh` appears to be a good solution
either.